+++ CHANGELOG.md
@@ -2,6 +2,39 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.2.11] - 2025-04-25
+
+### 🐛 Bug Fixes
+
+- Tune gossip params (#433)
+- Make the sleep in consensus critical path async (#436)
+- Read nodes only attempt sync against connected peers (#446)
+
+## [0.2.10] - 2025-04-23
+
+### 🐛 Bug Fixes
+
+- Increase output port capacity so we don't lose messages (#426)
+- Increase sync timeout and let read nodes connect to more validators (#429)
+
+### ⚙️ Miscellaneous Tasks
+
+- Add logging for sync (#428)
+- Add more logging for sync (#430)
+
+## [0.2.9] - 2025-04-22
+
+### 🐛 Bug Fixes
+
+- Reduce read node sync timeout (#422)
+
+## [0.2.8] - 2025-04-22
+
+### 🐛 Bug Fixes
+
+- Increase gossip connection limits  (#413)
+- Stop dialing discovered peers by default (#419)
+
 ## [0.2.7] - 2025-04-22
 
 ### 🚀 Features